Lab 8 Completed, Needs comments

+++ Lab8Part3/TimerLib_part3.h
@@ -5,4 +5,4 @@ void MotorSpeed(double x);
 void Keypad_init(void);
 int Keypad_Read(void);
 void SysTick_ms_delay(int n);
-int Press_Convert(int n);
+double Press_Convert(int n);
